Don't strip the root fs at the compilation time. strip is now done at the deployment.

+++ init_sd.sh
@@ -21,9 +21,9 @@ fi
 
 source ${BASE_DIR}/configs/${TARGET_NAME}/config.sh || exit 1
 
-echo $1
-
-gcc scripts/fix_bin_paths.c -o scripts/fix_bin_paths
+##########################################################################
+# Duplicate root fs
+##########################################################################
 
 export TARGET_ROOTFS_MIRROR=${BASE_DIR}/targets/${TARGET_NAME}/fs_mirror
 
@@ -34,14 +34,48 @@ cd     ${TARGET_ROOTFS_MIRROR}
 
 cp -av ${TARGET_ROOTFS}/* .
 
-find . -type f -exec ${BASE_DIR}/scripts/fix_bin_paths {} ${TARGET_ROOTFS} \;
+##########################################################################
+# Strip the whole root fs !
+##########################################################################
+
+rm -R bin/ld bin/ld.bfd bin/as bin/objdump bin/strip bin/ar bin/ranlib bin/nm bin/readelf bin/omshell
+rm -R man
+rm -R ssl
+rm -R share/man
+rm -R share/doc
+rm -R share/gtk-doc
+rm -R share/locale/*
+rm -R share/info/*
+rm -R share/i18n
+rm -R include
+find ./lib -type f -name "*.la" -delete
+find ./lib -type f -name "*.a" -delete
+find ./lib64 -type f -name "*.la" -delete
+find ./lib64 -type f -name "*.a" -delete
+rm ./bin/smbd/*.old
+rm ./bin/smbclient/*.old
+
+find ./ -type f -print0  -print | xargs -0 ${TGT_MACH}-strip --strip-debug --strip-unneeded
+
+find ./bin   -type f -exec ${TGT_MACH}-strip --strip-debug --strip-unneeded {} \;
+find ./lib   -type f -exec ${TGT_MACH}-strip --strip-debug --strip-unneeded {} \;
+find ./lib64 -type f -exec ${TGT_MACH}-strip --strip-debug --strip-unneeded {} \;
+find ./sbin  -type f -exec ${TGT_MACH}-strip --strip-debug --strip-unneeded {} \;
+find ./usr   -type f -exec ${TGT_MACH}-strip --strip-debug --strip-unneeded {} \;
+
+# Fix buggy path...
+gcc scripts/fix_bin_paths.c -o scripts/fix_bin_paths
+
+find ./ -type f -exec ${BASE_DIR}/scripts/fix_bin_paths {} ${TARGET_ROOTFS} \;
+
+##########################################################################
+# Copy configs files and do some last fixes...
+##########################################################################
 
 chmod +x "./lib/libc.so" || exit 1
 chmod +x "./lib/libpthread.so" || exit 1
 
-cp -R ${BASE_DIR}/configs/${TARGET_NAME}/rootfs_cfg/bin/* ./bin/
-cp -R ${BASE_DIR}/configs/${TARGET_NAME}/rootfs_cfg/etc/* ./etc/
-cp -R ${BASE_DIR}/configs/${TARGET_NAME}/rootfs_cfg/usr/* ./usr/
+cp -R ${BASE_DIR}/configs/${TARGET_NAME}/rootfs_cfg/* ./
 
 chmod +x ./etc/init.d/rcS
 chmod +x ./usr/share/udhcpc/*
@@ -51,19 +85,9 @@ mkdir ramdisk
 mkdir mnt/tmp
 mkdir usr/share/empty
 
-chown -R root:root ./*
-
-#Strip
-rm -R bin/ld bin/ld.bfd bin/as bin/objdump bin/strip bin/ar bin/ranlib bin/nm bin/readelf bin/omshell
-rm -R man
-rm -R ssl
-rm -R share/man
-rm -R share/locale/*
-rm -R share/info/*
-rm -R share/i18n
-rm -R include
-
+##########################################################################
 # Post process install...
+##########################################################################
 
 if [ -f ${BASE_DIR}/configs/${TARGET_NAME}/install_post_process.sh ]
 then
@@ -73,26 +97,33 @@ then
 )
 fi
 
+##########################################################################
+# Copy to Flash media !
+##########################################################################
 
-# Copy to SD
 cd ..
 
-sudo umount $1
+if [ ! -f $1 ]; then
+    echo "Copy to Flash media ..."
 
-sudo mkfs.ext4 $1
+    sudo umount $1
 
-mkdir mount_point
+    sudo mkfs.ext4 $1
 
-sudo mount $1 mount_point || exit 1
+    mkdir mount_point
 
-cd mount_point
+    sudo mount $1 mount_point || exit 1
 
-sudo cp -av ${TARGET_ROOTFS_MIRROR}/* .
+    cd mount_point
 
-sudo chown -R root *
-sudo chgrp -R root *
-cd ..
+    sudo cp -av ${TARGET_ROOTFS_MIRROR}/* .
 
-sudo umount $1
+    sudo chown -R root *
+    sudo chgrp -R root *
+    cd ..
 
-echo done...
+    sudo umount $1
+fi
+
+echo  Done !
+
